requirements.txt
boto3
requests
beautifulsoup4
pdfminer.six
python-docx
scikit-learn